Cleanup - organize scheduler plugin by their functionality instead of type
