+++ pkg/epp/scheduling/plugins/kvcache/kvcache_scorer.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package scorer
+package kvcache
 
 import (
 	"sigs.k8s.io/gateway-api-inference-extension/pkg/epp/scheduling/types"
