+++ pkg/epp/scheduling/plugins/lora/lora_affinity_filter.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package filter
+package lora
 
 import (
 	"math/rand"
