+++ pkg/epp/scheduling/plugins/queue/queue_scorer_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package scorer
+package queue
 
 import (
 	"context"
